fix(supabaseClient): separate supabase clients by client & server + change type check
