+++ src/app/auth/confirm/route.ts
@@ -1,42 +1,16 @@
-import { createServerClient } from "@supabase/ssr";
 import { NextResponse } from "next/server";
-import { cookies } from "next/headers";
 import { EmailOtpType } from "@supabase/supabase-js";
+import { createClient } from "@/utils/supabase/server";
 
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url);
   const token_hash = searchParams.get("token_hash");
   const type = searchParams.get("type") as EmailOtpType;
   const next = searchParams.get("next");
-  const cookieStore = cookies();
-
-  if (
-    !process.env.NEXT_PUBLIC_SUPABASE_URL ||
-    !process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
-  ) {
-    throw new Error("Supabase URL and Anon Key must be defined");
-  }
-
-  const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY,
-    {
-      cookies: {
-        get(name) {
-          return cookieStore.get(name)?.value;
-        },
-        set(name, value, options) {
-          cookieStore.set({ name, value, ...options });
-        },
-        remove(name, options) {
-          cookieStore.set({ name, value: "", ...options });
-        },
-      },
-    }
-  );
+  const serverClient = createClient();
 
   if (token_hash && type && next) {
-    const { error } = await supabase.auth.verifyOtp({
+    const { error } = await serverClient.auth.verifyOtp({
       type,
       token_hash,
     });
