--- src/utils/supabaseClient.ts
@@ -1,16 +1,0 @@
-"use client";
-
-import { createBrowserClient } from "@supabase/ssr";
-if (
-  !process.env.NEXT_PUBLIC_SUPABASE_URL ||
-  !process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
-) {
-  throw new Error("Supabase URL and Anon Key must be defined");
-}
-
-const supabase = createBrowserClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL,
-  process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
-);
-
-export { supabase };
